Adds functions to get all header names
This adds `get_request_header_names` and `get_response_header_names`
functions, which iterate in consistent order.

The results are a concatenation of NUL-terminated header names. Ex
"Accept\0Host\0"
