+++ handler/mosn/go.mod
@@ -3,26 +3,19 @@ module github.com/http-wasm/http-wasm-host-go/handler/mosn
 go 1.18
 
 require (
-	github.com/envoyproxy/go-control-plane v0.10.0
 	github.com/http-wasm/http-wasm-host-go v0.0.0-20221011003337-81d8f3916f05
-	github.com/urfave/cli v1.22.1
 	mosn.io/api v1.1.0
 	mosn.io/mosn v1.1.0
 	mosn.io/pkg v1.1.0
 )
 
 require (
-	github.com/StackExchange/wmi v0.0.0-20190523213315-cbe66965904d // indirect
 	github.com/andybalholm/brotli v1.0.2 // indirect
 	github.com/antlr/antlr4 v0.0.0-20200503195918-621b933c7a7f // indirect
 	github.com/c2h5oh/datasize v0.0.0-20171227191756-4eba002a5eae // indirect
-	github.com/cncf/xds/go v0.0.0-20211001041855-01bcc9b48dfe // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/dchest/siphash v1.2.1 // indirect
-	github.com/envoyproxy/protoc-gen-validate v0.1.0 // indirect
 	github.com/ghodss/yaml v1.0.0 // indirect
-	github.com/go-ole/go-ole v1.2.4 // indirect
-	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/mock v1.6.0 // indirect
 	github.com/golang/protobuf v1.5.0 // indirect
 	github.com/google/cel-go v0.5.1 // indirect
@@ -39,10 +32,10 @@ require (
 	github.com/oklog/run v1.0.0 // indirect
 	github.com/rcrowley/go-metrics v0.0.0-20200313005456-10cdbea86bc0 // indirect
 	github.com/russross/blackfriday/v2 v2.0.1 // indirect
-	github.com/shirou/gopsutil v3.20.11+incompatible // indirect
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
 	github.com/tetratelabs/wazero v1.0.0-pre.2.0.20221003082636-0b4dbfd8d6ca // indirect
 	github.com/trainyao/go-maglev v0.0.0-20200611125015-4c1ae64d96a8 // indirect
+	github.com/urfave/cli v1.22.1 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.14.1-0.20200605121233-ac51d598dc54 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
@@ -57,8 +50,6 @@ require (
 	google.golang.org/protobuf v1.27.1 // indirect
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	istio.io/gogo-genproto v0.0.0-20210113155706-4daf5697332f // indirect
-	mosn.io/holmes v1.0.1 // indirect
 	mosn.io/proxy-wasm-go-host v0.1.1-0.20210524020952-3fb13ba763a6 // indirect
 )
